fix(photo-uploader): move accountSite query to SiteSelector component, and handle error when there is no siteId
